requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
playwright==1.44.0
python-binance==1.0.19
streamlit==1.32.0
typer==0.12.3
google-generativeai==0.7.2
python-dotenv==1.0.1
plotly==5.20.0
pandas==2.2.1
rich==13.7.1
